+++ CHANGELOG.md
@@ -6,7 +6,7 @@ __BREAKING CHANGES__
 
 * Require gdbots/pbj and gdbots/schemas v3.x.
 * Publish as type module in npm.
-* Add throw option to Pbjx.trigger, default to true to match php implementation.
+* Add throwEx option to Pbjx.trigger, default to true to match php implementation.
 
 
 ## v2.0.1